requirements.txt
streamlit
langchain
openai
pinecone-client
tiktoken